Repeater chapter and unify siunitx
I rewrote the text for the repeater appendix, please read through and see what you think.  Please also tell me if you think we should change the tables to state the frequencies from the POV of the repeater.  I think that is how the PTS question goes.

I also changed a few places all over to move away from explicit use of \Omega for ohm and changed some units that were written out as mA, mW, etc to use siunitx.  This is just a preparatory step towards moving the version of siunitx.